src/test/kotlin/ru/romanow/ValidPalindromeTest.kt
package ru.romanow

import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.extension.ExtensionContext
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.Arguments
import org.junit.jupiter.params.provider.Arguments.of
import org.junit.jupiter.params.provider.ArgumentsProvider
import org.junit.jupiter.params.provider.ArgumentsSource
import java.util.stream.Stream

class ValidPalindromeTest {

    @ArgumentsSource(ValueProvider::class)
    @ParameterizedTest(name = "#{index} – Check string {0}")
    fun isPalindrome(
        s: String,
        result: Boolean,
    ) {
        val obj = ValidPalindrome()
        assertThat(obj.isPalindrome(s)).isEqualTo(result)
    }

    internal class ValueProvider : ArgumentsProvider {
        override fun provideArguments(context: ExtensionContext): Stream<Arguments> =
            Stream.of(
                of("A man, a plan, a canal: Panama", true),
                of("race a car", false),
                of(" ", true),
            )
    }
}
